requirements.txt
# The Goodies Smart Home System - Python Requirements

# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Authentication & Security
argon2-cffi>=23.1.0
PyJWT>=2.8.0
qrcode[pil]>=7.4.2
cryptography>=41.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
hypothesis>=6.90.0

# Development tools
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0

# API documentation
httpx>=0.25.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0

# Chat Interface & LLM Support
llama-cpp-python>=0.3.0
diskcache>=5.6.0